move an item test to its own file
